Module ibc_chain_registry::formatter

Contains traits to format the URL of API endpoints from a &str to any type. Contains struct to build a tendermint_rpc::Url representing a WebSocket URL from a RPC URL and to parse or build a valid http::Uri from an (in)complete GRPC URL.

Structs

SimpleGrpcFormatter

SimpleGrpcFormatter contains methods to parse or build a valid http::Uri from an (in)complete GRPC URL.

SimpleWebSocketFormatter

SimpleWebSocketFormatter contains methods to parse a valid WebSocket URL from an RPC URL.

Traits

UriFormatter

UriFormatter contains the basic expectations to parse a valid URL from a &str.