ibc_chain_registry::formatter
Trait UriFormatter

pub trait UriFormatter {
    type OutputFormat;

    // Required method
    fn parse_or_build_address(
        url: &str,
    ) -> Result<Self::OutputFormat, RegistryError>;
}

UriFormatter contains the basic expectations to parse a valid URL from a &str.

Required Associated Types

type OutputFormat

Expected output format of the formatter.

Required Methods

fn parse_or_build_address( url: &str, ) -> Result<Self::OutputFormat, RegistryError>

Attempts to parse the given input as a OutputFormat. If the parsed URL is not complete, this method attempts to fill in the necessary missing pieces. Returns a RegistryError if this attempt fails.

Arguments

url - A string slice that holds the url which should be formatted

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.