Module ibc_chain_registry::querier

Contains traits to query nodes of a given chain from their APIs. Contains struct to perform a health check on a gRPC/WebSocket endpoint and to retrieve the max_block_size from a RPC endpoint.

Structs

GrpcHealthCheckQuerier

GrpcHealthCheckQuerier connects to a list of gRPC endpoints and returns the URL of the first one to answer.

HermesConfigData

Data which must be retrieved from RPC endpoints for Hermes

SimpleHermesRpcQuerier

SimpleHermesRpcQuerier retrieves HermesConfigData by querying a list of RPC endpoints through their WebSocket API and returns the result of the first endpoint to answer.

Traits

QueryContext

QueryContext represents the basic expectations for a query

QueryTypes

QueryTypes represents the basic types required to query a node