Trait ibc_chain_registry::querier::QueryContext

pub trait QueryContext: QueryTypes {
    fn query_error(chain_name: String) -> Self::QueryError;
    fn query<'async_trait>(
        url: Self::QueryInput
    ) -> Pin<Box<dyn Future<Output = Result<Self::QueryOutput, Self::QueryError>> + Send + 'async_trait>>
    where
        Self: 'async_trait;

    fn query_healthy<'async_trait>(
        chain_name: String,
        urls: Vec<Self::QueryInput>
    ) -> Pin<Box<dyn Future<Output = Result<Self::QueryOutput, Self::QueryError>> + Send + 'async_trait>>
    where
        Self: Send + 'async_trait,
    { ... }
}

QueryContext represents the basic expectations for a query

Required Methods

fn query_error(chain_name: String) -> Self::QueryError

Return an error specific to the query which is retured when query_healthy fails

Arguments

chain_name - A string slice that holds the name of a chain

fn query<'async_trait>(
url: Self::QueryInput
) -> Pin<Box<dyn Future<Output = Result<Self::QueryOutput, Self::QueryError>> + Send + 'async_trait>>where
Self: 'async_trait,

Query an endpoint and return the result

Arguments

url - A QueryInput object that holds the data needed to query a node

Provided Methods

fn query_healthy<'async_trait>(
chain_name: String,
urls: Vec<Self::QueryInput>
) -> Pin<Box<dyn Future<Output = Result<Self::QueryOutput, Self::QueryError>> + Send + 'async_trait>>where
Self: Send + 'async_trait,

Query every endpoint from a list of urls and return the output of the first one to answer.

Arguments

chain_name - A string that holds the name of a chain

urls - A vector of urls to query