ibc_chain_registry::querier
Trait QueryTypes

pub trait QueryTypes {
    type QueryInput: Debug + Send;
    type QueryOutput;
    type QueryError;
}

QueryTypes represents the basic types required to query a node

Required Associated Types

type QueryInput: Debug + Send

QueryInput represents the data needed to query a node. It is typically a URL

type QueryOutput

QueryOutput represents the data returned by your query

type QueryError

QueryOutput represents the error returned when a query fails