Crate ibc_chain_registry

Modules

asset_list

Contains models for serializing and deserializing assets.json for a given chain originally from https://github.com/PeggyJV/ocular/blob/main/ocular/src/registry/assets.rs

chain

Contains models for serializing and deserializing chain.json for a given chain Taken from https://github.com/PeggyJV/ocular/blob/main/ocular/src/registry/chain.rs

constants

error

fetchable

Contains the trait required to fetch and deserialize data from the chain repository

formatter

Contains traits to format the URL of API endpoints from a &str to any type. Contains struct to build a tendermint_rpc::Url representing a WebSocket URL from a RPC URL and to parse or build a valid http::Uri from an (in)complete GRPC URL.

paths

querier

Contains traits to query nodes of a given chain from their APIs. Contains struct to perform a health check on a gRPC/WebSocket endpoint and to retrieve the max_block_size from a RPC endpoint.