Trait ibc_chain_registry::formatter::UriFormatter

pub trait UriFormatter {
    type OutputFormat;
    fn parse_or_build_address(
        url: &str
    ) -> Result<Self::OutputFormat, RegistryError>;
}

UriFormatter contains the basic expectations to parse a valid URL from a &str.

Associated Types

type OutputFormat

Expected output format of the formatter.

Required methods

fn parse_or_build_address(
url: &str
) -> Result<Self::OutputFormat, RegistryError>

Attempts to parse the given input as a OutputFormat. If the parsed URL is not complete, this method attempts to fill in the necessary missing pieces. Returns a RegistryError if this attempt fails.

Arguments

url - A string slice that holds the url which should be formatted